+++ packages/webview_flutter/lib/platform_interface.dart
@@ -182,6 +182,14 @@ abstract class WebViewPlatformController {
         "WebView loadUrl is not implemented on the current platform");
   }
 
+  /// Loads the specified html.
+  ///
+  /// `html` must not be null.
+  Future<void> loadHtml(String html) {
+    throw UnimplementedError(
+        "WebView loadHtml is not implemented on the current platform");
+  }
+
   /// Updates the webview settings.
   ///
   /// Any non null field in `settings` will be set as the new setting value.
@@ -427,6 +435,8 @@ class CreationParams {
   /// The `autoMediaPlaybackPolicy` parameter must not be null.
   CreationParams({
     this.initialUrl,
+    this.html,
+    this.baseUrl,
     this.webSettings,
     this.javascriptChannelNames,
     this.userAgent,
@@ -438,6 +448,12 @@ class CreationParams {
   ///
   /// When null the webview will be created without loading any page.
   final String initialUrl;
+
+  /// The initial baseUrl to load in the webview. It will work with html only.
+  final String baseUrl;
+
+  /// The initial html to load in the webview. This will be preferred over initialUrl if it exists.
+  final String html;
 
   /// The initial [WebSettings] for the new webview.
   ///