+++ packages/webview_flutter/lib/webview_flutter.dart
@@ -144,6 +144,8 @@ class WebView extends StatefulWidget {
     Key key,
     this.onWebViewCreated,
     this.initialUrl,
+    this.html,
+    this.baseUrl,
     this.javascriptMode = JavascriptMode.disabled,
     this.javascriptChannels,
     this.navigationDelegate,
@@ -209,6 +211,12 @@ class WebView extends StatefulWidget {
 
   /// The initial URL to load.
   final String initialUrl;
+
+  /// The initial baseURL to load.
+  final String baseUrl;
+
+  /// The initial static html to load.
+  final String html;
 
   /// Whether Javascript execution is enabled.
   final JavascriptMode javascriptMode;
@@ -389,6 +397,8 @@ class _WebViewState extends State<WebView> {
 CreationParams _creationParamsfromWidget(WebView widget) {
   return CreationParams(
     initialUrl: widget.initialUrl,
+    html: widget.html,
+    baseUrl: widget.baseUrl,
     webSettings: _webSettingsFromWidget(widget),
     javascriptChannelNames: _extractChannelNames(widget.javascriptChannels),
     userAgent: widget.userAgent,
@@ -546,6 +556,15 @@ class WebViewController {
     assert(url != null);
     _validateUrlString(url);
     return _webViewPlatformController.loadUrl(url, headers);
+  }
+
+  /// Loads the specified html.
+  ///
+  /// `html` must not be null.
+  ///
+  Future<void> loadHtml(String html) async {
+    assert(html != null);
+    return _webViewPlatformController.loadHtml(html);
   }
 
   /// Accessor to the current URL that the WebView is displaying.